add Enum for image handling and restructure the Shared project
