+++ BlazingTrails.Api/Features/ManageTrails/UploadTrailImageEndpoint.cs
@@ -1,6 +1,6 @@
 using Ardalis.ApiEndpoints;
 using BlazingTrails.Api.Persistence;
-using BlazingTrails.Shared.Features.ManageTrails;
+using BlazingTrails.Shared.Features.ManageTrails.Shared;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using SixLabors.ImageSharp;
