+++ BlazingTrails.Client/Features/ManageTrails/AddTrail/AddTrailHandler.cs
@@ -1,5 +1,5 @@
 using System.Net.Http.Json;
-using BlazingTrails.Shared.Features.ManageTrails;
+using BlazingTrails.Shared.Features.ManageTrails.AddTrail;
 using MediatR;
 
 namespace BlazingTrails.Client.Features.ManageTrails.AddTrail;
