+++ BlazingTrails.Client/Features/ManageTrails/AddTrail/AddTrailPage.razor
@@ -1,4 +1,6 @@
 @page "/addTrail"
+@using BlazingTrails.Shared.Features.ManageTrails.AddTrail
+@using BlazingTrails.Shared.Features.ManageTrails.Shared
 @inject IMediator Mediator
 
 <PageTitle>Add Trail - Blazing Trails</PageTitle>
