+++ BlazingTrails.Client/Features/ManageTrails/Shared/TrailForm.razor
@@ -1,4 +1,6 @@
 @using BlazingTrails.Client.Validation
+@using BlazingTrails.Shared.Features.ManageTrails.Shared
+
 <EditForm EditContext="_editContext" OnValidSubmit="SubmitForm">
     <FluentValidationValidator/>
     <FormSection Title="Basic Details" HelpText="This information is used to identify the trail and can be searched to help hikers find it.">
