+++ BlazingTrails.Client/Features/ManageTrails/Shared/UploadTrailImageHandler.cs
@@ -1,4 +1,4 @@
-using BlazingTrails.Shared.Features.ManageTrails;
+using BlazingTrails.Shared.Features.ManageTrails.Shared;
 using MediatR;
 
 namespace BlazingTrails.Client.Features.ManageTrails.Shared;
