+++ BlazingTrails.Shared/BlazingTrails.Shared.csproj
@@ -13,4 +13,8 @@
     <PackageReference Include="Microsoft.AspNetCore.Components.Web" Version="8.0.4" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Features\ManageTrails\EditTrail\" />
+  </ItemGroup>
+
 </Project>
