--- BlazingTrails.Shared/Features/ManageTrails/RouteInstructionValidator.cs
@@ -1,12 +1,0 @@
-using FluentValidation;
-
-namespace BlazingTrails.Shared.Features.ManageTrails;
-
-public class RouteInstructionValidator : AbstractValidator<TrailDto.RouteInstruction>
-{
-    public RouteInstructionValidator()
-    {
-        RuleFor(x => x.Stage).NotEmpty().WithMessage("Please enter a stage");
-        RuleFor(x => x.Description).NotEmpty().WithMessage("Please enter a description");
-    }
-}
