+++ BlazingTrails.Shared/Features/ManageTrails/Shared/ImageAction.cs
@@ -1,0 +1,8 @@
+namespace BlazingTrails.Shared.Features.ManageTrails.Shared;
+
+public enum ImageAction
+{
+    None,
+    Add,
+    Remove
+}
