+++ BlazingTrails.Shared/Features/ManageTrails/Shared/UploadTrailImageRequest.cs
@@ -1,7 +1,7 @@
 using MediatR;
 using Microsoft.AspNetCore.Components.Forms;
 
-namespace BlazingTrails.Shared.Features.ManageTrails;
+namespace BlazingTrails.Shared.Features.ManageTrails.Shared;
 
 public record UploadTrailImageRequest(int TrailId, IBrowserFile File) : IRequest<UploadTrailImageRequest.Response>
 {
